fix(ops): the profile endpoint answered 500 to every call it ever received
Found by reading, which is what I should have done three runs ago instead of
waiting for a log that was never going to name it.

`GET /users/{username}/ops-profile` computes thirteen figures in one row.
Twelve are `count(*)`, which PostgreSQL returns as bigint, and the struct
reads all thirteen as `i64`. The thirteenth went through `date_part` and was
cast `::INT`, making it the one int4 in the row. sqlx does not widen an
integer to fit: it refused to decode the row, and the endpoint returned
DATABASE_ERROR to every request since it was written.

The suite reported one failure, and I read that as one flaky test. Exactly
one test reaches this endpoint, and only on its way to checking that a lapsed
certification does not score — so a wholly broken endpoint and a single
awkward test looked identical from the outside.

That is the part worth fixing beyond the cast: 404 of 922 registered routes
are never called by any test. Writing 404 tests is not the answer, but the
hole that hid this one is now closed — a smoke test asserts that no ops read
endpoint answers 5xx. It checks the shape of the answer, not its content:
unauthenticated is 401, absent is 404, refused is 403, and 500 is never
right. It fails the day a query stops decoding, which is the failure that
hid.

The other explicit `::INT` casts in the codebase were checked and match the
Rust types that read them.

+++ src/services/ops_profile.rs
@@ -160,8 +160,16 @@ async fn measure(db: &PgPool, user_id: Uuid) -> Result<Measurements, AppError> {
                 AND g.domain = 'ops')
                 AS review_grid_average,
 
+            -- BIGINT, not INT. Every other figure here is a `count(*)`, which
+            -- PostgreSQL returns as bigint, and the struct reads them all as
+            -- `i64`. This one went through `date_part`, so the cast decided
+            -- the column's width — and `::INT` made it the one int4 in the
+            -- row. sqlx does not widen: it refused to decode the whole row,
+            -- so `GET /users/{username}/ops-profile` answered 500 to every
+            -- call it has ever received. One test reaches this endpoint,
+            -- which is why one test failed rather than the suite.
             (SELECT COALESCE(
-                        date_part('year', age(NOW(), min(a.issued_at)))::INT + 1,
+                        date_part('year', age(NOW(), min(a.issued_at)))::BIGINT + 1,
                         0)
                FROM attestations a
               WHERE a.user_id = $1 AND a.revoked_at IS NULL
